Cocus Challenge for Marcus Bergo
First of all, thanks for the opportunity. So bacically I went trough using a Makefile design to ease the setup and clarify my intentions. I wrote some unittests, for good manners :). Moreover, I plot an example of the output (output_example.txt) and I had some problems with IAM which I places on problems-and-solutions.md.